+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,40 @@
+package roomescape.controller;
+
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.service.ReservationService;
+
+@RestController
+@RequestMapping("/reservations")
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Reservation>> findAll() {
+        return ResponseEntity.ok().body(reservationService.findReservations());
+    }
+
+    @PostMapping
+    public ResponseEntity<Reservation> create(@RequestBody ReservationRequest reservationRequest) {
+        return ResponseEntity.ok().body(reservationService.addReservation(reservationRequest));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationService.deleteReservation(id);
+        return ResponseEntity.ok().build();
+    }
+}
